+++ tests/Integration/Balikobot/GetAddServiceOptionsMethodTest.php
@@ -2,7 +2,6 @@
 
 namespace Inspirum\Balikobot\Tests\Integration\Balikobot;
 
-use Inspirum\Balikobot\Contracts\ExceptionInterface;
 use Inspirum\Balikobot\Definitions\ServiceType;
 use Inspirum\Balikobot\Definitions\Shipper;
 
@@ -12,12 +11,15 @@ class GetAddServiceOptionsMethodTest extends AbstractBalikobotTestCase
     {
         $service = $this->newBalikobot();
 
-        $options = $service->getAddServiceOptions(Shipper::CP, ServiceType::CP_CE);
+        $options = $service->getAddServiceOptions(Shipper::CP);
 
         $this->assertTrue(count($options) > 0);
-        foreach ($options as $code => $option) {
-            $this->assertTrue(is_int($code));
-            $this->assertTrue(is_string($option));
+        foreach ($options as $serviceType => $serviceOptions) {
+            $this->assertTrue(is_string($serviceType));
+            foreach ($serviceOptions as $code => $option) {
+                $this->assertTrue(is_int($code) || is_string($code));
+                $this->assertTrue(is_string($option));
+            }
         }
     }
 
@@ -35,6 +37,23 @@ class GetAddServiceOptionsMethodTest extends AbstractBalikobotTestCase
     }
 
     public function testValidRequestWithFullData()
+    {
+        $service = $this->newBalikobot();
+
+        $options = $service->getAddServiceOptions(Shipper::CP, null, true);
+
+        $this->assertTrue(count($options) > 0);
+        foreach ($options as $serviceType => $serviceOptions) {
+            $this->assertTrue(is_string($serviceType));
+            foreach ($serviceOptions as $code => $option) {
+                $this->assertTrue(is_array($option));
+                $this->assertTrue(is_string($option['name']));
+                $this->assertTrue($option['code'] === (string) $code);
+            }
+        }
+    }
+
+    public function testValidRequestWithServiceWithFullData()
     {
         $service = $this->newBalikobot();